requirements.txt
paramiko>=4.0.0
pyyaml==6.0.1
click==8.1.7
